+++ addons/fray/src/state_management/combat_situation_builder.gd
@@ -1,5 +1,6 @@
 extends Reference
 
+const State = preload("res://addons/fray/lib/state_machine/state.gd")
 const CombatSituation = preload("combat_situation.gd")
 const CombatState = preload("combat_state.gd")
 const InputCondition = preload("transitions/conditions/input_condition.gd")
@@ -29,6 +30,7 @@ var _global_builder_by_state: Dictionary
 ## Type: Dictionary<String, String[]>
 var _transition_rules: Dictionary
 
+var _behavior_state: State
 var _func_new_button: FuncRef
 var _func_new_sequence: FuncRef
 
@@ -46,6 +48,9 @@ func _init() -> void:
 ## Returns a newly constructed CombatSituation
 func build(initial_state: String) -> CombatSituation:
 	var cs := CombatSituation.new()
+
+	cs.behavior_state = _behavior_state
+
 	for state_name in _state_by_name:
 		cs.add_state(state_name, _state_by_name[state_name])
 
@@ -87,6 +92,7 @@ func build(initial_state: String) -> CombatSituation:
 	_builder_by_state_tuple.clear()
 	_global_builder_by_state.clear()
 	_transition_rules.clear()
+	_behavior_state = null
 	return cs
 
 ## Adds a new state to the situation.
@@ -105,6 +111,13 @@ func set_state(state_name: String, state: CombatState) -> Reference:
 	else:
 		push_warning("Failed to set state. State with name '%s' does not exist." % state_name)
 	return self
+
+## Sets behavior state used by situation
+## This is a completely optional configuration.
+func set_behavior(behavior_state: State) -> Reference:
+	_behavior_state = behavior_state
+	return self
+
 
 ## Creates a new transition from one state to another.
 ## If states does not already exist they will be created.